requirements.txt
openai
python-dotenv>=1.0.0
dash>=2.16.0
dash-bootstrap-components>=1.5.0
sounddevice>=0.4.6
scipy>=1.10.0
numpy>=1.23.0
pydub 
simpleaudio